首页新版改版

+++ zigbee-app-new/src/main/webapp/app/js/rentRdc.js
@@ -125,26 +125,44 @@ $().ready(function () {
             if(rdc.istemperaturestandard==1){
                 approve='<i class="iconfont orange">&#xe63b;</i><i class="orange">未认证</i><i class="iconfont orange">&#xe6e9;</i><i class="orange">冷链委温度达标库</i>'
             }else{
-                approve='<i class="iconfont orange">&#xe63b;</i><i class="orange">未认证</i>'
+                approve='<i class="iconfont orange">&#xe63b;</i><i class="orange">未认证</i>';
             }
         };
+        var collectWords='<a class="fr noCollect" onclick="collect(this,'+rdc.id+')"><i class="iconfont">&#xe605;</i><em>收藏</em></a>';
+        if(rdc.collectUserIds && window.user){
+            for(var i=0;i<rdc.collectUserIds.length;i++){
+                if(rdc.collectUserIds[i]==window.user.id){
+                    collectWords='<a class="fr hasCollect" onclick="collect(this,'+rdc.id+')"><i class="iconfont">&#xe60c;</i><em>已收藏</em></a>';
+                }
+            }
+        }
         var score = [
             '<li class="imgCell"><a href="storehousedetail.html?id='+rdc.id+'"  onclick="getSoll()"><span><img src="'+rdc.logo+'" alt=""></span><div>'+
             '<p class="ellipsis">'+rdc.title+'</p><em>信息完整度<i class="blue">'+rdc.infoIntegrity+'</i></em><p class="position omg">'+approve+'</p>'+
             '<p class="grab orange">'+rdc.unitPrice+'<br><span>元/㎡/天</span></p></div><div class="flex"><div class="item"><h4>'+rdc.sqm+'㎡</h4>'+
             '<p>总面积</p></div><div class="item"><h4>'+rdc.sqm+'㎡</h4><p>可租面积</p></div><div class="item"><h4 class="omg">'+rdc.detlAddress+'</h4><p>地址</p></div></div></a>'+
             '<div class="btnFn clearfix"><a href="storehousedetail.html?id='+rdc.id+'" class="fl"><i class="iconfont">&#xe65b;</i>查看</a>'+
-            '<a class="fr noCollect" onclick="collect(this)"><i class="iconfont">&#xe605;</i><em>收藏</em></a><a class="fr"><i class="iconfont">&#xe66c;</i>咨询</a></div></li>'
+            collectWords+'<a class="fr"><i class="iconfont">&#xe66c;</i>咨询</a></div></li>'
         ];
         return score.join("");
     }
-    collect=function(ops) {
+    collect=function(ops,id) {
+        if(!(window.user && window.user.id!=0)){
+            layer.open({content: "请登入后收藏！", btn: '确定'});
+            return;
+        }
         var em = $(ops);
         if(em.hasClass('noCollect')){
+            $.post(ER.root+"/i/collect/addCollectRdc",{uid:window.user.id,collectId:id,collectType:2},function (data) {
+
+            });
             em.removeClass('noCollect').addClass('hasCollect');
             em.children('i').html('&#xe60c;');
             em.children('em').html('已收藏');
         }else{
+            $.post(ER.root+"/i/collect/delCollectById",{uid:window.user.id,collectId:id,collectType:2},function (data) {
+
+            });
             em.addClass('noCollect').removeClass('hasCollect');
             em.children('i').html('&#xe605;');
             em.children('em').html('收藏');
